Build the core RV64 proving path and tighten shared-bus semantics
This commit lands the prover/runtime work needed to make RV64 a real proving target inside Nightstream instead of a synthetic alias over the RV32 path.

Main changes:
- add a first-class RV64 trace/CCS foundation in neo-memory, including RV64 layouts, witness extraction, CCS wiring, proof profile validation, ELF loading metadata, lowering scaffolding, memory remap support, and dedicated RV64 width-sidecar support
- add the RV64 shard path in neo-fold, including RV64 RAM bridging, exact register-output linkage claims, RV64-specific Route-A width/final-output checks, and a split rv64_trace_shard module so the top-level file stays within the project file-size constraint
- extend the RISC-V lookup/decomposition stack for RV64 W-operations (Mulw/Divuw/Remuw/Divw/Remw) through helper-backed virtual decomposition rather than pretending those operations are owned by direct architectural lookup tables
- widen the shout-key transport to carry exact RV64 lookup keys, fixing the previous u64 truncation bug in the active path
- tighten shared-bus and Route-A verification logic so prover/verifier opening sets, exact-register linkage, and RV64 width ownership stay aligned
- improve low-level proving kernels and supporting math paths that the new RV64 route relies on, including Ajtai decomposition reuse, SuperNEO cached evaluation reuse, and related bridge/session plumbing
- update the integration, shared-bus, trace-twist, trace-shout, and RV64-focused tests to cover the new semantics and to lock in the invariants around exact transport and helper-owned W-op decomposition

Why this grouping:
- these files together define the underlying proving/runtime semantics
- the follow-up commit will add the real RV64IM note guest/runtime harness and canonical real-ELF repro tests on top of this base
- keeping the substrate separate from the concrete note harness makes the history easier to audit and bisect

Validation already run against this change-set before the fresh-eyes cleanup:
- cargo test -p neo-fold --release --test test_rv64_trace_wiring_prove_verify -- --nocapture
- cargo test -p neo-fold --release --features poseidon-precompile --test test_rv64_note_from_elf -- --ignored --nocapture --test-threads=1
- cargo test --workspace --release

Fresh-eyes cleanup included in this commit:
- extracted rv64_trace_shard helpers into a dedicated submodule to satisfy AGENTS.md file-size limits
- removed panic/unwrap-style access in the exact-register debug verification path where checked errors were more appropriate
- replaced a remaining repo-root expect in the RV64 guest helper path with a normal Result-based error

+++ crates/neo-memory/tests/riscv_rv32m_masked_columns.rs
@@ -102,6 +102,6 @@ fn rv32_trace_shout_event_table_mulh_key_matches_operands() {
         .find(|row| row.opcode == Some(RiscvOpcode::Mulh))
         .expect("expected MULH row");
 
-    let expected_key = interleave_bits(/*lhs=*/ 3, /*rhs=*/ 5) as u64;
+    let expected_key = interleave_bits(/*lhs=*/ 3, /*rhs=*/ 5);
     assert_eq!(mulh_row.key, expected_key, "MULH key must encode rs1/rs2 values");
 }
